day_2/deploy_image_classifier_docker/requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cycler==0.10.0
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
starlette==0.14.2
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
wheel==0.36.2